openai/pyproject.toml
[project]
# Renamed from "openai" to avoid clashing with the upstream "openai" package dependency.
name = "openai-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.3.2",
    "python-dotenv>=1.1.1",
    "sounddevice>=0.5.2",
    "websockets>=15.0.1",
    "openai>=1.40.0",
    "gradio>=4.44.0",
    "websocket-client>=1.8.0",
]
